move find methods to conversations
